improve(Battle): code to avoid double-checks
The previous solution to avoid checking a cell twice involved changing
the contents of the cell in the board to a different character (kind of
cheating actually). It also was rather expensive because for each
detected boat cell the whole line had to be changed (since the lines are
represented by immutable javascript strings).

Much simpler and efficient is to simply accumulate the coordinates of
found boats and then skip the inner loop when the coordinates are
already in the accumulator.